requirements.txt
boto3==1.26.96
python-dotenv==1.0.0
